refactor(mimo-v2.5): B-class structural cleanups
- Make the embed runner truly model-agnostic: move the MiMo audio placeholder
  contract out of EmbedModelRunner into the host codec (MiMoV25AudioCodecProcessor.
  validate_audio_placeholder_contract, numpy-only) and expose it via a generic
  model hook (MiMoV2_5Embedding.validate_embed_inputs); the runner now dispatches
  via getattr(model, 'validate_embed_inputs') with no MiMo names. Drop the now-unused
  math import from the runner. Relocate the 3 contract tests to call the codec method.
- Collapse near-duplicate _encode_image/_encode_video into one _encode_visual(kind).
- Centralize weight mappings: move create_mimo_vision_weight_mappings + CONV3D_TORCH_
  TO_JAX into weights_mapping.py next to the text/audio builders (vision_encoder's
  standalone loader + embedding import it from there).

Skipped from the B list with rationale: config-fallback dedup (deduping 10 trivial
lines needs a new leaf module since host-side processor must not import the jax
model stack); audio_codec file split (torch is already lazy-imported, benefit
marginal vs churn); initialize_jit eager rename and vision standalone-loader removal
(both touch shared/alignment paths that can't be verified in this env).

14 test modules / 83 tests green.
