+++ python/sgl_jax/srt/multimodal/models/mimo_v2_5/audio_codec_processor.py
@@ -328,6 +328,63 @@ class MiMoV25AudioCodecProcessor:
             )
 
     @staticmethod
+    def validate_audio_placeholder_contract(input_ids, omni_inputs, audio_codes=None) -> None:
+        """Host-side guard: audio codes must have a scatter target and a placeholder
+        count that matches the audio mm_items, else the embed scatter silently drops
+        the audio (review R2-6). No-op for requests without audio codes.
+        """
+        if input_ids is None or audio_codes is None or not isinstance(omni_inputs, dict):
+            return
+
+        audio_token_id = omni_inputs.get("audio_token_id")
+        if audio_token_id is None:
+            raise ValueError(
+                "MiMo-V2.5 received audio codes but omni_inputs has no audio_token_id; "
+                "cannot place audio embeddings. Check that the model config / processor "
+                "resolves audio_token_id (expected 151669)."
+            )
+
+        items = omni_inputs.get("mm_items", []) or []
+        expected = 0
+        for raw_item in items:
+            if isinstance(raw_item, dict):
+                modality_name = str(raw_item.get("modality")).lower()
+                is_audio = modality_name == "audio" or modality_name.endswith("audio")
+                meta = raw_item.get("model_specific_data") or {}
+                feature = raw_item.get("feature")
+            else:
+                is_audio = raw_item.is_audio() if hasattr(raw_item, "is_audio") else False
+                meta = getattr(raw_item, "model_specific_data", None) or {}
+                feature = getattr(raw_item, "feature", None)
+            if not is_audio:
+                continue
+            if not (meta.get("is_codes") or "token_lengths" in meta):
+                continue
+            token_lengths = meta.get("token_lengths")
+            if token_lengths is not None:
+                expected += sum(int(length) for length in token_lengths)
+                continue
+            group_size = int(meta.get("group_size") or 1)
+            if feature is None:
+                continue
+            shape = getattr(feature, "shape", ())
+            if len(shape) == 2:
+                expected += math.ceil(int(shape[0]) / group_size)
+
+        if expected == 0:
+            return
+
+        ids = input_ids.tolist() if hasattr(input_ids, "tolist") else list(input_ids)
+        actual = sum(1 for token in ids if int(token) == int(audio_token_id))
+        if actual != expected:
+            raise ValueError(
+                "MiMo-V2.5 audio placeholder count mismatch before embed stage: "
+                f"found {actual} audio pad tokens, expected {expected} from mm_items "
+                f"token_lengths. input_ids_len={len(ids)} audio_codes_shape="
+                f"{getattr(audio_codes, 'shape', None)}"
+            )
+
+    @staticmethod
     def validate_placeholder_count(
         input_ids: list[int] | np.ndarray | None,
         payload: MiMoV25AudioPayload | None,
